dev-2.0-sprint-2: fixed tomcat_init_manager to read the correct values from tomcat-users.xml fixed tomcat installer to search for pre-existing admin user in tomcat-users.xml file

+++ installers/TomcatLinuxInstaller/src/files/setup.sh
@@ -44,7 +44,10 @@ fi
 cd $TOMCAT_CONF
 mv tomcat-users.xml tomcat-users.xml.old
 # XXX TODO: cleanup tomcat users, admin user should not be assigned both manager-gui and manager-script roles
-sed 's/<\/tomcat-users>/\n  <role rolename="manager-gui"\/>\n  <role rolename="manager"\/>\n  <user username="'$WEBSERVICE_USERNAME'" password="'$WEBSERVICE_PASSWORD'" roles="manager,manager-gui,manager-script"\/>\n<\/tomcat-users>/g' tomcat-users.xml.old > tomcat-users.xml
+userExists=`grep "username=\"$WEBSERVICE_USERNAME\"" tomcat-users.xml`
+if [ -z "$userExists" ]; then
+  sed 's/<\/tomcat-users>/\n  <role rolename="manager-gui"\/>\n  <role rolename="manager"\/>\n  <user username="'$WEBSERVICE_USERNAME'" password="'$WEBSERVICE_PASSWORD'" roles="manager,manager-gui,manager-script"\/>\n<\/tomcat-users>/g' tomcat-users.xml.old > tomcat-users.xml
+fi
 rm  -f tomcat-users.xml.old
